update admin routes
